Closes the residual commit c0d7ebe5 deliberately left open, but with the opposite conclusion from that commit's own speculative note. The movsx in our build traced to baseline_travel being declared int while its AAS_AreaTravelTimeToGoalArea call two lines below -- are short.
c0d7ebe5 had guessed the fix belonged in AAS_AreaTravelTimeToGoalArea
itself (widen its __int16 return to int, since Q3's cognate returns
plain int). Checking the actual disasm first disproved that: every
-- the 16-bit return is genuine, matching Q3 would have regressed a
byte-identical function. Q3 widened the signature as it evolved
(added travelflags, origin -> vec3_t) and apparently the return type
with it; that's real divergence, not a decompiler artifact here.

Retyping baseline_travel to short instead (matching its siblings)
Whole-image re-audit: 724 -> 725 MATCH, zero collateral elsewhere.

Source-commit: 1b1b68aee9cce96fde90ca2b9a0e57445337911d

+++ botlib/botlib.c
@@ -13459,7 +13459,7 @@ int __cdecl AAS_AlternativeRouteGoals(
 {
   int   startareanum;
   int   goalareanum;
-  int   baseline_travel;
+  short baseline_travel;
   short travel_to_start;
   short travel_to_goal;
   int   areanum;
